Setup docker orchestration for testing
The contains the runtime for the library and testing its example webapp.

This includes a change to `(start-webapp!)`, such that it respects the
PORT env var. This is because docker-compose will expose a fixed port
and then pass that in with PORT. Typing in 8080 every time would be a
right pain.