Add in-mem user store
This commit defines an interface for user stores and implements it in
memory for something to test with.

The set of tests defined here should be common for each store added so
we can be sure there is semantic parity. It is unlikely they are perfect
so these will grow but should be abstract of any one store.