Add reset password action
When resetting your password you first off enter an email address, then
receive a URL in an email and then user this email to pick a new
password and reset it.

I considered putting the picking of a password in the first stage to
drop a page, but a survey of a few websites showed picking after email
was more common and I don't want to confuse anyone. Testing should be
done.

We can also confirm the user if they reset their password before getting
round to confirming. This flow confirms their email address as they take
action on something sent to it, so it must be valid if they complete.

I had to change the structure of the IUserStore protocol as it assumed
that we had a user map when generating a reset token. This is not the
case, all we have is an email, so we switch it to lookup a user based on
an email address and then proceed as before.